added authorization at handler level to world and persona

+++ src/main/java/me/moirai/discordbot/core/application/usecase/persona/DeletePersonaHandler.java
@@ -3,9 +3,11 @@ package me.moirai.discordbot.core.application.usecase.persona;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 import me.moirai.discordbot.common.annotation.UseCaseHandler;
+import me.moirai.discordbot.common.exception.AssetAccessDeniedException;
 import me.moirai.discordbot.common.exception.AssetNotFoundException;
 import me.moirai.discordbot.common.usecases.AbstractUseCaseHandler;
 import me.moirai.discordbot.core.application.usecase.persona.request.DeletePersona;
+import me.moirai.discordbot.core.domain.persona.Persona;
 import me.moirai.discordbot.core.domain.persona.PersonaRepository;
 
 @UseCaseHandler
@@ -13,6 +15,7 @@ public class DeletePersonaHandler extends AbstractUseCaseHandler<DeletePersona, 
 
     private static final String PERSONA_NOT_FOUND = "Persona was not found";
     private static final String ID_CANNOT_BE_NULL_OR_EMPTY = "Persona ID cannot be null or empty";
+    private static final String USER_NO_PERMISSION_IN_PERSONA = "User does not have permission to delete the persona";
 
     private final PersonaRepository repository;
 
@@ -32,8 +35,12 @@ public class DeletePersonaHandler extends AbstractUseCaseHandler<DeletePersona, 
     @Override
     public Void execute(DeletePersona request) {
 
-        repository.findById(request.getId())
+        Persona persona = repository.findById(request.getId())
                 .orElseThrow(() -> new AssetNotFoundException(PERSONA_NOT_FOUND));
+
+        if (!persona.canUserWrite(request.getRequesterDiscordId())) {
+            throw new AssetAccessDeniedException(USER_NO_PERMISSION_IN_PERSONA);
+        }
 
         repository.deleteById(request.getId());
