+++ src/main/java/me/moirai/discordbot/core/application/usecase/world/DeleteWorldHandler.java
@@ -3,9 +3,11 @@ package me.moirai.discordbot.core.application.usecase.world;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 import me.moirai.discordbot.common.annotation.UseCaseHandler;
+import me.moirai.discordbot.common.exception.AssetAccessDeniedException;
 import me.moirai.discordbot.common.exception.AssetNotFoundException;
 import me.moirai.discordbot.common.usecases.AbstractUseCaseHandler;
 import me.moirai.discordbot.core.application.usecase.world.request.DeleteWorld;
+import me.moirai.discordbot.core.domain.world.World;
 import me.moirai.discordbot.core.domain.world.WorldRepository;
 
 @UseCaseHandler
@@ -13,6 +15,7 @@ public class DeleteWorldHandler extends AbstractUseCaseHandler<DeleteWorld, Void
 
     private static final String WORLD_TO_BE_VIEWED_WAS_NOT_FOUND = "World to be viewed was not found";
     private static final String ID_CANNOT_BE_NULL_OR_EMPTY = "World ID cannot be null or empty";
+    private static final String USER_NO_PERMISSION_IN_PERSONA = "User does not have permission to delete the persona";
 
     private final WorldRepository repository;
 
@@ -31,8 +34,12 @@ public class DeleteWorldHandler extends AbstractUseCaseHandler<DeleteWorld, Void
     @Override
     public Void execute(DeleteWorld command) {
 
-        repository.findById(command.getId())
+        World world = repository.findById(command.getId())
                 .orElseThrow(() -> new AssetNotFoundException(WORLD_TO_BE_VIEWED_WAS_NOT_FOUND));
+
+        if (!world.canUserWrite(command.getRequesterDiscordId())) {
+            throw new AssetAccessDeniedException(USER_NO_PERMISSION_IN_PERSONA);
+        }
 
         repository.deleteById(command.getId());
